Implement inheritance in Nokia model and add basic tests
The Nokia model was modified to inherit from the Smartphone abstract
class and the InstallApplication method was overridden to tailor its
functionality to Nokia phones. Created a new object of the Nokia class
in Program.cs file and ran basic tests to validate its behavior. These
changes aimed to maintain object-oriented practices and ensure that the
Nokia model behaves as a specific type of smartphone.